ble_py_server/requirements.txt
kafka-python==2.0.2
PyJWT==2.8.0
pydbus==0.6.0